refactor: List-related elements move Router.jsx->ListContext.jsx
